+++ src/components/VideoMosaic/VideoMosaic.component.jsx
@@ -1,23 +1,8 @@
 import { Card, CardActionArea, CardMedia, Grid, Typography } from '@material-ui/core';
 import React from 'react';
 import { useHistory } from 'react-router-dom';
+import { shortenDescription, shortenTitle } from '../../utils/shortenText';
 import { CustomCardContent, CustomCardContentTitle } from './VideoMosaic.styled';
-
-export function shortenTitle(title) {
-  if (title.length > 40) {
-    const croppedTitle = title.substring(0, 39);
-    return `${croppedTitle}...`;
-  }
-  return title;
-}
-
-export function shortenDescription(description) {
-  if (description.length > 150) {
-    const croppedDescription = description.substring(0, 149);
-    return `${croppedDescription}...`;
-  }
-  return description;
-}
 
 function VideoMosaic(props) {
   const history = useHistory();
