+++ src/utils/theme.js
@@ -1,0 +1,9 @@
+export const darkTheme = {
+  palette: { normal: '#263238', dark: '#000a12', light: '#4f5b62' },
+  fontColor: '#FFFFFF',
+};
+
+export const lightTheme = {
+  palette: { normal: '#FFFFFF', dark: '#CCCCCC', light: '#FFFFFF' },
+  fontColor: '#263238',
+};
